requested changes done reset head, rename file and remove lock file
