+++ modules/districtmDmxBidAdapter.md
